pyproject.toml
[project]
name = "stock-screener"
version = "0.1.0"
description = "GARP stock screener using yfinance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yfinance>=0.2.49",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "fpdf2>=2.7.0",
    "ruff>=0.14.8",
    "isort>=7.0.0",
    "scipy>=1.15.3",
    "pyyaml>=6.0.3",
]
